refactor: rename Media folder to media
