+++ app/routes/movies/$movieId/index.tsx
@@ -5,7 +5,7 @@ import { Text, Row, Col, Spacer, Divider } from '@nextui-org/react';
 import { useRouteData } from 'remix-utils';
 import { getSimilar, getVideos, getCredits, getRecommendation } from '~/services/tmdb/tmdb.server';
 import { IMovieDetail, ICast, ICrew } from '~/services/tmdb/tmdb.types';
-import MediaList from '~/src/components/Media/MediaList';
+import MediaList from '~/src/components/media/MediaList';
 import PeopleList from '~/src/components/people/PeopleList';
 import useMediaQuery from '~/hooks/useMediaQuery';
