+++ app/routes/tv-shows/$tvId.tsx
@@ -4,7 +4,7 @@ import { useCatch, useLoaderData, Outlet, Link, RouteMatch } from '@remix-run/re
 import { Container } from '@nextui-org/react';
 
 import { getTvShowDetail } from '~/services/tmdb/tmdb.server';
-import MediaDetail from '~/src/components/Media/MediaDetail';
+import MediaDetail from '~/src/components/media/MediaDetail';
 import CatchBoundaryView from '~/src/components/CatchBoundaryView';
 import ErrorBoundaryView from '~/src/components/ErrorBoundaryView';
 import i18next from '~/i18n/i18next.server';
