+++ app/src/components/media/MediaDetail.tsx
